+++ libs/error-tracking/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @protolabsai/error-tracking
 
+## 0.53.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @protolabsai/types@0.94.0
+  - @protolabsai/utils@0.94.0
+
 ## 0.53.40
 
 ### Patch Changes